+++ packages/dialect-react-ui/components/Chat/screens/Main/ThreadsList/MessagePreview.tsx
@@ -2,15 +2,45 @@ import React from 'react';
 import { useApi, DialectAccount, formatTimestamp } from '@dialectlabs/react';
 import { display } from '@dialectlabs/web3';
 import Avatar from '../../../../Avatar';
+import clsx from 'clsx';
 
 type PropsType = {
   dialect: DialectAccount;
   onClick: () => void;
+  disabled?: boolean;
 };
+
+function FirstMessage({ dialect }: { dialect: DialectAccount }) {
+  const { wallet } = useApi();
+  const messages = dialect.dialect.messages || [];
+
+  if (dialect.dialect.encrypted) {
+    return (
+      <div className="dt-text-sm dt-opacity-30 dt-italic dt-mb-2">
+        Encrypted message
+      </div>
+    );
+  }
+
+  return messages && messages?.length > 0 ? (
+    <div className="dt-max-w-full dt-text-sm dt-opacity-50 dt-mb-2 dt-truncate">
+      <span className="dt-opacity-50">
+        {messages[0].owner.toString() === wallet?.publicKey?.toString() &&
+          'You:'}
+      </span>{' '}
+      {messages[0].text}
+    </div>
+  ) : (
+    <div className="dt-text-sm dt-opacity-30 dt-italic dt-mb-2">
+      No messages yet
+    </div>
+  );
+}
 
 export default function MessagePreview({
   dialect,
   onClick,
+  disabled = false,
 }: PropsType): JSX.Element {
   const { wallet } = useApi();
   const otherMembers = dialect?.dialect.members.filter(
@@ -21,12 +51,13 @@ export default function MessagePreview({
   );
   const otherMemberStr = otherMembersStrs[0];
 
-  const messages = dialect.dialect.messages || [];
-
   return (
     <div
-      className="dt-flex dt-space-x-2 dt-items-center dt-w-full dt-cursor-pointer"
-      onClick={onClick}
+      className={clsx(
+        disabled ? 'dt-cursor-not-allowed' : 'dt-cursor-pointer',
+        'dt-flex dt-space-x-2 dt-items-center dt-w-full'
+      )}
+      onClick={!disabled ? onClick : undefined}
     >
       <div className="dt-flex">
         <Avatar publicKey={otherMembers[0].publicKey} size="regular" />
@@ -34,19 +65,7 @@ export default function MessagePreview({
       <div className="dt-flex dt-grow dt-border-b dt-border-neutral-600 dt-justify-between dt-truncate dt-pr-2">
         <div className="dt-flex dt-flex-col dt-max-w-full dt-truncate">
           {dialect?.dialect.members.length > 0 && <div>{otherMemberStr}</div>}
-          {messages && messages?.length > 0 ? (
-            <div className="dt-max-w-full dt-text-sm dt-opacity-50 dt-mb-2 dt-truncate">
-              <span className="dt-opacity-50">
-                {messages[0].owner.toString() ===
-                  wallet?.publicKey?.toString() && 'You:'}
-              </span>{' '}
-              {messages[0].text}
-            </div>
-          ) : (
-            <div className="dt-text-sm dt-opacity-30 dt-italic dt-mb-2">
-              No messages yet
-            </div>
-          )}
+          <FirstMessage dialect={dialect} />
         </div>
         <div className="dt-text-xs dt-opacity-30">
           {formatTimestamp(dialect.dialect.lastMessageTimestamp)}