+++ packages/dialect-react-ui/components/Icon/X.tsx
@@ -4,9 +4,9 @@ import { SVGProps } from 'react';
 const SvgX = (props: SVGProps<SVGSVGElement>) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
+    className="x_svg__h-6 x_svg__w-6"
     fill="none"
-    height={24}
-    width={24}
+    viewBox="0 0 24 24"
     stroke="currentColor"
     {...props}
   >