+++ packages/dialect-react-ui/components/Notifications/index.tsx
@@ -14,6 +14,7 @@ import {
   Centered,
   Divider,
   Footer,
+  NetworkBadge,
   useBalance,
   ValueRow,
 } from '../common';
@@ -79,29 +80,6 @@ function Header(props: {
       </div>
       <Divider className="dt-mx-2" />
     </>
-  );
-}
-
-function NetworkBadge({ network = 'devnet' }: { network?: string | null }) {
-  const { textStyles, colors } = useTheme();
-  let color = 'dt-text-green-600';
-  if (network === 'devnet') {
-    color = 'dt-text-yellow-600';
-  }
-  if (network === 'localnet') {
-    color = 'dt-text-red-600';
-  }
-  return (
-    <span
-      className={cs(
-        'dt-py-0.5 dt-px-1 dt-rounded-sm',
-        textStyles.small,
-        colors.highlight,
-        color
-      )}
-    >
-      {network}
-    </span>
   );
 }